Started adding browser information to C# invocations
Created the IBrowserInvocation interface to start collecting more information about an action. Currently have coords of the item, screen dimension & browser information being passed up. Also in this commit are more changes on the docs, I wanted to render the views as partials to make it easier to work on them however that isn't behaving too well. I also have put showdown.js into the web app so I can render markdown as its a bit easier to write the documentation with 😏